race-weekend-api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy>=2.0
psycopg2-binary
passlib[bcrypt]
PyJWT
redis>=5.0
httpx>=0.24
pytest
pytest-asyncio
pytest-cov
email-validator
